verificationhelper: split QR code tests into separate file
